Working on Employees/Companies Management
Trying to make it easy to redirect from companies to employees (to see
the employees in the specific company). Right now it works, but we're
just using the template to display the employees correctly. We can't
interact with anything in that template, since we're still in the
companies folder with the companies index.php.